Integrate PSO into existing WebSocket infra; clean up comments and dead code
- Move PSOOptimizationConsumer into existing apps/core/consumers.py (next to TaskStatusConsumer)

- Remove parallel apps/core/websocket/ package; register PSO route in existing apps/core/routing.py

- Remove dev-status/phase narration comments and debug prints across plate girder backend files

- Delete unused PSODashboard and its ParallelCoordinatesPlot/PerformanceMap/CrossSectionPreview components (superseded by OptimizationGraph)

- Reuses existing Celery app for the PSO task (no parallel infrastructure)
