+++ backend/apps/core/routing.py
@@ -1,9 +1,8 @@
-from django.urls import path, re_path
+﻿from django.urls import path, re_path
 from . import consumers
-from .websocket import consumers as pso_consumers
 
 websocket_urlpatterns = [
     path('ws/tasks/<str:task_id>/', consumers.TaskStatusConsumer.as_asgi()),
     # Real-time PSO optimization for Plate Girder
-    re_path(r'^ws/optimize/plate-girder/$', pso_consumers.PSOOptimizationConsumer.as_asgi()),
+    re_path(r'^ws/optimize/plate-girder/$', consumers.PSOOptimizationConsumer.as_asgi()),
 ]
