--- backend/apps/core/websocket/__init__.py
@@ -1,2 +1,0 @@
-# WebSocket package for real-time PSO optimization updates
-
