+++ backend/apps/modules/flexure_member/submodules/plate_girder/pso_imports.py
@@ -1,28 +1,3 @@
-"""
-Core PSO Algorithms and Section Utilities - Import Module
-
-This module centralizes imports of PSO algorithms and section utilities from osdag_core.
-These will be used in the Celery task (tasks.py) for optimization.
-
-All classes and functions are already implemented in osdag_core, we just import and use them.
-"""
-
-# ==============================================================================
-# 2.1 Global Best PSO
-# ==============================================================================
-# Source: osdag_core/design_type/plate_girder/core/pso_optimizer.py
-# Features:
-# - Standard PSO velocity update: V = w*V + c1*r1*(P_best - X) + c2*r2*(G_best - X)
-# - Constraint-aware initialization (only feasible particles)
-# - Constraint checking during position updates
-# - Resampling if particle becomes infeasible
-# - Progress callback support (emits iteration, particle index, position, cost)
-# Parameters:
-# - n_particles = 50
-# - iterations = 100
-# - w = 0.4 (inertia weight)
-# - c1 = 1.5 (cognitive coefficient)
-# - c2 = 1.5 (social coefficient)
 try:
     from osdag_core.design_type.plate_girder.core.pso_optimizer import GlobalBestPSO
     GLOBAL_BEST_PSO_AVAILABLE = True
@@ -31,17 +6,6 @@ except ImportError as e:
     GlobalBestPSO = None
     print(f"Warning: Could not import GlobalBestPSO: {e}")
 
-
-# ==============================================================================
-# 2.2 Intelligent PSO
-# ==============================================================================
-# Source: osdag_core/design_type/plate_girder/optimization/intelligent_pso.py
-# Features:
-# - Discrete variable snapping (to standard thicknesses)
-# - Smart boundary clamping (inelastic collision at boundaries)
-# - Continuous search space with discrete evaluation
-# - Soft constraint handling (penalties)
-# Discrete Variables: Standard thicknesses for tw, tf, t_stiff (6, 8, 10, 12, 16, 20, 25, 32, 40 mm)
 try:
     from osdag_core.design_type.plate_girder.optimization.intelligent_pso import IntelligentPSO
     INTELLIGENT_PSO_AVAILABLE = True
@@ -50,15 +14,6 @@ except ImportError as e:
     IntelligentPSO = None
     print(f"Warning: Could not import IntelligentPSO: {e}")
 
-
-# ==============================================================================
-# 2.3 Section Utilities
-# ==============================================================================
-# Source: osdag_core/design_type/plate_girder/core/section.py
-# Functions:
-# - calc_yj(): Calculate yj for unsymmetric sections (IS 800:2007 E.3.2.2)
-# - classify_section(): Classify plate girder section (Plastic/Compact/Semi-Compact/Slender)
-# - shear_stress_unsym_I(): Calculate shear stress distribution
 try:
     from osdag_core.design_type.plate_girder.core.section import (
         calc_yj,
@@ -74,16 +29,7 @@ except ImportError as e:
     print(f"Warning: Could not import section utilities: {e}")
 
 
-# ==============================================================================
-# Verification Function
-# ==============================================================================
 def verify_imports():
-    """
-    Verify that all required PSO algorithms and section utilities can be imported.
-    
-    Returns:
-        dict: Status of each import with availability flags
-    """
     status = {
         'global_best_pso': {
             'available': GLOBAL_BEST_PSO_AVAILABLE,
@@ -105,9 +51,6 @@ def verify_imports():
     return status
 
 
-# ==============================================================================
-# Export all imports for use in tasks.py and other modules
-# ==============================================================================
 __all__ = [
     'GlobalBestPSO',
     'IntelligentPSO',
@@ -119,4 +62,3 @@ __all__ = [
     'SECTION_UTILITIES_AVAILABLE',
     'verify_imports'
 ]
-
